Switch to logic for LUTs
The RAM thing for LUTs didn't make much sense.
